sig: reconcile BubbleHelp and RES_OpenFileFromVolumePtr signatures
text.h's BubbleHelp declaration now matches its definition
(int *rect, char *text, int font) instead of (unsigned int*, unsigned int,
unsigned int), clearing the three C4028. Callers updated to match: DAT_004b9f78
typed int[4], DAT_007fe048 typed char* (it is freed and holds a string), and
title's rc is int[4] with the (unsigned int) cast on the text arg dropped.

resource.c RES_OpenFileFromVolumePtr passes entry->name (array decay to char*)
instead of &entry->name, clearing the C4047/C4024 at line 532.

All affected functions stay at their prior match; aggregate unchanged at 63.08%.

+++ src/legoland/title.c
@@ -755,7 +755,7 @@ void FUN_00491080(char *param_1, int param_2, int param_3, int param_4, int para
 
 // FUNCTION: LEGOLAND 0x004910f0
 void FUN_004910f0(void) {
-    unsigned int rc[4];
+    int rc[4];
     char **line;
     int y;
     int n;
@@ -782,7 +782,7 @@ void FUN_004910f0(void) {
         if ((int)GetTicks() > (int)DAT_00798888) {
             DAT_00798888 = 0;
         }
-        BubbleHelp(rc, (unsigned int)((char **)&DAT_007cb140)[DAT_00798884], 2);
+        BubbleHelp(rc, ((char **)&DAT_007cb140)[DAT_00798884], 2);
     }
 }
